services/server/requirements.txt
Flask>=2.0
psycopg2-binary
celery>=5.0
redis>=4.0
pandas>=1.0
numpy>=1.20
scikit-learn>=1.0
requests>=2.20 # Bao gồm cho module crawler hoặc các API gọi ra ngoài